Possible refactor for TaskTable.jsx
