+++ client/app/queue/components/TaskTable.jsx
@@ -17,237 +17,20 @@ import PropTypes from 'prop-types';
 
 import QueueTable from '../QueueTable';
 import Checkbox from '../../components/Checkbox';
-import DocketTypeBadge from '../../components/DocketTypeBadge';
-import HearingBadge from './HearingBadge';
-import OnHoldLabel, { numDaysOnHold } from './OnHoldLabel';
-import ReaderLink from '../ReaderLink';
-import CaseDetailsLink from '../CaseDetailsLink';
-import ContinuousProgressBar from '../../components/ContinuousProgressBar';
 
 import { setSelectionOfTaskOfUser } from '../QueueActions';
-import { renderAppealType, taskHasCompletedHold, actionNameOfTask, regionalOfficeCity } from '../utils';
+
 import { DateString } from '../../util/DateUtil';
-import {
-  CATEGORIES,
-  redText,
-  LEGACY_APPEAL_TYPES,
-  DOCKET_NAME_FILTERS
-} from '../constants';
+
 import COPY from '../../../COPY.json';
-import CO_LOCATED_ADMIN_ACTIONS from '../../../constants/CO_LOCATED_ADMIN_ACTIONS.json';
 import QUEUE_CONFIG from '../../../constants/QUEUE_CONFIG.json';
+import * as TaskColumns from './TaskTableCols';
 
 const hasDASRecord = (task, requireDasRecord) => {
   return (task.appeal.isLegacyAppeal && requireDasRecord) ? Boolean(task.taskId) : true;
 };
 
-const collapseColumn = (requireDasRecord) => (task) => hasDASRecord(task, requireDasRecord) ? 1 : 0;
-
-export const docketNumberColumn = (tasks, requireDasRecord) => {
-  return {
-    header: COPY.CASE_LIST_TABLE_DOCKET_NUMBER_COLUMN_TITLE,
-    name: QUEUE_CONFIG.DOCKET_NUMBER_COLUMN,
-    enableFilter: true,
-    tableData: tasks,
-    columnName: 'appeal.docketName',
-    customFilterLabels: DOCKET_NAME_FILTERS,
-    anyFiltersAreSet: true,
-    label: 'Filter by docket name',
-    valueName: 'docketName',
-    backendCanSort: true,
-    valueFunction: (task) => {
-      if (!hasDASRecord(task, requireDasRecord)) {
-        return null;
-      }
-
-      return <React.Fragment>
-        <DocketTypeBadge name={task.appeal.docketName} number={task.appeal.docketNumber} />
-        <span>{task.appeal.docketNumber}</span>
-      </React.Fragment>;
-    },
-    span: collapseColumn(requireDasRecord),
-    getSortValue: (task) => {
-      if (!hasDASRecord(task, requireDasRecord)) {
-        return null;
-      }
-
-      return `${task.appeal.docketName || ''} ${task.appeal.docketNumber}`;
-    }
-  };
-};
-
-export const hearingBadgeColumn = () => {
-  return {
-    header: '',
-    name: QUEUE_CONFIG.HEARING_BADGE_COLUMN,
-    valueFunction: (task) => <HearingBadge task={task} />
-  };
-};
-
-export const detailsColumn = (tasks, requireDasRecord, userRole) => {
-  return {
-    header: COPY.CASE_LIST_TABLE_VETERAN_NAME_COLUMN_TITLE,
-    name: QUEUE_CONFIG.CASE_DETAILS_LINK_COLUMN,
-    valueFunction: (task) => <CaseDetailsLink
-      task={task}
-      appeal={task.appeal}
-      userRole={userRole}
-      disabled={!hasDASRecord(task, requireDasRecord)} />,
-    backendCanSort: true,
-    getSortValue: (task) => {
-      const vetName = task.appeal.veteranFullName.split(' ');
-      // only take last, first names. ignore middle names/initials
-
-      return `${_.last(vetName)} ${vetName[0]}`;
-    }
-  };
-};
-
-export const taskColumn = (tasks) => {
-  return {
-    header: COPY.CASE_LIST_TABLE_TASKS_COLUMN_TITLE,
-    name: QUEUE_CONFIG.TASK_TYPE_COLUMN,
-    enableFilter: true,
-    tableData: tasks,
-    columnName: 'label',
-    anyFiltersAreSet: true,
-    customFilterLabels: CO_LOCATED_ADMIN_ACTIONS,
-    label: 'Filter by task',
-    valueName: 'label',
-    valueFunction: (task) => actionNameOfTask(task),
-    backendCanSort: true,
-    getSortValue: (task) => actionNameOfTask(task)
-  };
-};
-
-export const regionalOfficeColumn = (tasks) => {
-  return {
-    header: COPY.CASE_LIST_TABLE_REGIONAL_OFFICE_COLUMN_TITLE,
-    name: QUEUE_CONFIG.REGIONAL_OFFICE_COLUMN,
-    enableFilter: true,
-    tableData: tasks,
-    columnName: 'closestRegionalOffice.location_hash.city',
-    anyFiltersAreSet: true,
-    label: 'Filter by regional office',
-    backendCanSort: true,
-    valueFunction: (task) => {
-      return regionalOfficeCity(task, true);
-    },
-    getSortValue: (task) => regionalOfficeCity(task)
-  };
-};
-
-export const issueCountColumn = (requireDasRecord) => {
-  return {
-    header: COPY.CASE_LIST_TABLE_APPEAL_ISSUE_COUNT_COLUMN_TITLE,
-    name: QUEUE_CONFIG.ISSUE_COUNT_COLUMN,
-    valueFunction: (task) => hasDASRecord(task, requireDasRecord) ? task.appeal.issueCount : null,
-    span: collapseColumn(requireDasRecord),
-    backendCanSort: true,
-    getSortValue: (task) => hasDASRecord(task, requireDasRecord) ? task.appeal.issueCount : null
-  };
-};
-
-export const typeColumn = (tasks, requireDasRecord) => {
-  return {
-    header: COPY.CASE_LIST_TABLE_APPEAL_TYPE_COLUMN_TITLE,
-    name: QUEUE_CONFIG.APPEAL_TYPE_COLUMN,
-    enableFilter: true,
-    tableData: tasks,
-    columnName: 'appeal.caseType',
-    anyFiltersAreSet: true,
-    label: 'Filter by type',
-    valueName: 'caseType',
-    valueFunction: (task) => hasDASRecord(task, requireDasRecord) ?
-      renderAppealType(task.appeal) :
-      <span {...redText}>{COPY.ATTORNEY_QUEUE_TABLE_TASK_NEEDS_ASSIGNMENT_ERROR_MESSAGE}</span>,
-    span: (task) => hasDASRecord(task, requireDasRecord) ? 1 : 5,
-    getSortValue: (task) => {
-      // We append a * before the docket number if it's a priority case since * comes before
-      // numbers in sort order, this forces these cases to the top of the sort.
-      if (task.appeal.isAdvancedOnDocket || task.appeal.caseType === LEGACY_APPEAL_TYPES.CAVC_REMAND) {
-        return `*${task.appeal.docketNumber}`;
-      }
-
-      return task.appeal.docketNumber;
-    }
-  };
-};
-
-export const assignedToColumn = (tasks) => {
-  return {
-    header: COPY.CASE_LIST_TABLE_APPEAL_LOCATION_COLUMN_TITLE,
-    name: QUEUE_CONFIG.TASK_ASSIGNEE_COLUMN,
-    enableFilter: true,
-    tableData: tasks,
-    columnName: 'assignedTo.name',
-    anyFiltersAreSet: true,
-    label: 'Filter by assignee',
-    valueFunction: (task) => task.assignedTo.name,
-    getSortValue: (task) => task.assignedTo.name
-  };
-};
-
-export const readerLinkColumn = (requireDasRecord, includeNewDocsIcon) => {
-  return {
-    header: COPY.CASE_LIST_TABLE_APPEAL_DOCUMENT_COUNT_COLUMN_TITLE,
-    name: QUEUE_CONFIG.DOCUMENT_COUNT_READER_LINK_COLUMN,
-    span: collapseColumn(requireDasRecord),
-    valueFunction: (task) => {
-      if (!hasDASRecord(task, requireDasRecord)) {
-        return null;
-      }
-
-      return <ReaderLink appealId={task.externalAppealId}
-        analyticsSource={CATEGORIES.QUEUE_TABLE}
-        redirectUrl={window.location.pathname}
-        appeal={task.appeal}
-        newDocsIcon={includeNewDocsIcon}
-        task={task}
-        docCountBelowLink />;
-    }
-  };
-};
-
-export const daysWaitingColumn = (requireDasRecord) => {
-  return {
-    header: COPY.CASE_LIST_TABLE_TASK_DAYS_WAITING_COLUMN_TITLE,
-    name: QUEUE_CONFIG.DAYS_WAITING_COLUMN,
-    span: collapseColumn(requireDasRecord),
-    tooltip: <React.Fragment>Calendar days since <br /> this case was assigned</React.Fragment>,
-    align: 'center',
-    valueFunction: (task) => {
-      return <React.Fragment>
-        <span className={taskHasCompletedHold(task) ? 'cf-red-text' : ''}>{moment().startOf('day').
-          diff(moment(task.assignedOn), 'days')}</span>
-        { taskHasCompletedHold(task) ? <ContinuousProgressBar level={moment().startOf('day').
-          diff(task.placedOnHoldAt, 'days')} limit={task.onHoldDuration} warning /> : null }
-      </React.Fragment>;
-    },
-    backendCanSort: true,
-    getSortValue: (task) => moment().startOf('day').
-      diff(moment(task.assignedOn), 'days')
-  };
-};
-
-export const daysOnHoldColumn = (requireDasRecord) => {
-  return {
-    header: COPY.CASE_LIST_TABLE_TASK_DAYS_ON_HOLD_COLUMN_TITLE,
-    name: QUEUE_CONFIG.TASK_HOLD_LENGTH_COLUMN,
-    span: collapseColumn(requireDasRecord),
-    tooltip: <React.Fragment>Calendar days since <br /> this case was placed on hold</React.Fragment>,
-    align: 'center',
-    valueFunction: (task) => {
-      return <React.Fragment>
-        <OnHoldLabel task={task} />
-        <ContinuousProgressBar limit={task.onHoldDuration} level={moment().startOf('day').
-          diff(task.placedOnHoldAt, 'days')} />
-      </React.Fragment>;
-    },
-    backendCanSort: true,
-    getSortValue: (task) => numDaysOnHold(task)
-  };
-};
+// const collapseColumn = (requireDasRecord) => (task) => hasDASRecord(task, requireDasRecord) ? 1 : 0;
 
 export class TaskTableUnconnected extends React.PureComponent {
   getKeyForRow = (rowNumber, object) => object.uniqueId
@@ -269,7 +52,7 @@ export class TaskTableUnconnected extends React.PureComponent {
   collapseColumnIfNoDASRecord = (task) => this.taskHasDASRecord(task) ? 1 : 0
 
   caseHearingColumn = () => {
-    return this.props.includeHearingBadge ? hearingBadgeColumn() : null;
+    return this.props.includeHearingBadge ? TaskColumns.hearingBadgeColumn() : null;
   }
 
   caseSelectColumn = () => {
@@ -289,12 +72,12 @@ export class TaskTableUnconnected extends React.PureComponent {
 
   caseDetailsColumn = () => {
     return this.props.includeDetailsLink ?
-      detailsColumn(this.props.tasks, this.props.requireDasRecord, this.props.userRole) :
+      TaskColumns.detailsColumn(this.props.tasks, this.props.requireDasRecord, this.props.userRole) :
       null;
   }
 
   caseTaskColumn = () => {
-    return this.props.includeTask ? taskColumn(this.props.tasks) : null;
+    return this.props.includeTask ? TaskColumns.taskColumn(this.props.tasks) : null;
   }
 
   caseDocumentIdColumn = () => {
@@ -318,20 +101,22 @@ export class TaskTableUnconnected extends React.PureComponent {
   }
 
   caseTypeColumn = () => {
-    return this.props.includeType ? typeColumn(this.props.tasks, this.props.requireDasRecord) : null;
+    return this.props.includeType ? TaskColumns.typeColumn(this.props.tasks, this.props.requireDasRecord) : null;
   }
 
   caseAssignedToColumn = () => {
-    return this.props.includeAssignedTo ? assignedToColumn(this.props.tasks) : null;
+    return this.props.includeAssignedTo ? TaskColumns.assignedToColumn(this.props.tasks) : null;
   }
 
   caseDocketNumberColumn = () => {
-    return this.props.includeDocketNumber ? docketNumberColumn(this.props.tasks, this.props.requireDasRecord) : null;
+    return this.props.includeDocketNumber ? TaskColumns.docketNumberColumn(this.props.tasks, this.props.requireDasRecord) : null;
   }
 
   caseIssueCountColumn = () => {
-    return this.props.includeIssueCount ? issueCountColumn(this.props.requireDasRecord) : null;
+    return this.props.includeIssueCount ? TaskColumns.issueCountColumn(this.props.requireDasRecord) : null;
   }
+
+  // valueFunction could be made into its own utility/component esp, if it's just repsonsible for returning cell value
 
   caseDueDateColumn = () => {
     return this.props.includeDueDate ? {
@@ -359,11 +144,11 @@ export class TaskTableUnconnected extends React.PureComponent {
   }
 
   caseDaysWaitingColumn = () => {
-    return this.props.includeDaysWaiting ? daysWaitingColumn(this.props.requireDasRecord) : null;
+    return this.props.includeDaysWaiting ? TaskColumns.daysWaitingColumn(this.props.requireDasRecord) : null;
   }
 
   caseDaysOnHoldColumn = () => {
-    return this.props.includeDaysOnHold ? daysOnHoldColumn(this.props.requireDasRecord) : null;
+    return this.props.includeDaysOnHold ? TaskColumns.daysOnHoldColumn(this.props.requireDasRecord) : null;
   }
 
   completedDateColumn = () => {
@@ -388,12 +173,12 @@ export class TaskTableUnconnected extends React.PureComponent {
 
   caseReaderLinkColumn = () => {
     return !this.props.userIsVsoEmployee && this.props.includeReaderLink ?
-      readerLinkColumn(this.props.requireDasRecord, this.props.includeNewDocsIcon) :
+      TaskColumns.readerLinkColumn(this.props.requireDasRecord, this.props.includeNewDocsIcon) :
       null;
   }
 
   caseRegionalOfficeColumn = () => {
-    return this.props.includeRegionalOffice ? regionalOfficeColumn(this.props.tasks) : null;
+    return this.props.includeRegionalOffice ? TaskColumns.regionalOfficeColumn(this.props.tasks) : null;
   }
 
   getQueueColumns = () =>
